remove pillow version to rely on recommended
Rely on pillows default version for a given interpreter. 7 doesn't (easily) compile with Python 3.9, which blocks initialization of the extension.
This is a bigger than expected issue as it can break/cause problems with ANY python workspace (ie not ones intended for use with Device Simulator) due to the default "always try to install requirements" extension setting.